+++ grpc_extra/decorators.py
@@ -15,18 +15,18 @@ from .constants import (
     GRPC_SERVICE_META,
 )
 from .ordering import BaseOrdering, get_default_ordering_class, resolve_ordering_class
-from .permissions import PermissionLike, resolve_permissions
 from .pagination import (
     BasePagination,
     get_default_pagination_class,
     resolve_pagination_class,
 )
+from .permissions import PermissionLike, resolve_permissions
+from .registry import MethodMeta, ServiceMeta, registry
 from .searching import (
     BaseSearching,
     get_default_searching_class,
     resolve_searching_class,
 )
-from .registry import MethodMeta, ServiceMeta, registry
 from .utils import to_upper_camel_case